Improved resolution of emulator
