+++ src/tooling/piral-cli/src/common/index.ts
@@ -36,3 +36,4 @@ export * from './template';
 export * from './url';
 export * from './version';
 export * from './watcher';
+export * from './website';
